+++ appalerts/rules.py
@@ -1,0 +1,21 @@
+"""Rules for the appalerts app"""
+
+import rules
+
+
+# Predicates -------------------------------------------------------------
+
+
+# None needed right now
+
+
+# Rules ------------------------------------------------------------------
+
+
+# Rules should not be needed, use permissions for user rights
+
+
+# Permissions ------------------------------------------------------------
+
+# Allow viewing alerts
+rules.add_perm('appalerts.view_alerts', rules.is_authenticated)