+++ appalerts/urls.py
@@ -1,0 +1,23 @@
+from django.conf.urls import url
+
+from appalerts import views, views_ajax
+
+app_name = 'appalerts'
+
+urlpatterns = [
+    url(
+        regex=r'^list$',
+        view=views.AppAlertListView.as_view(),
+        name='list',
+    ),
+    url(
+        regex=r'^redirect/(?P<appalert>[0-9a-f-]+)$',
+        view=views.AppAlertLinkRedirectView.as_view(),
+        name='redirect',
+    ),
+    url(
+        regex=r'^dismiss/(?P<appalert>[0-9a-f-]+)$',
+        view=views_ajax.AppAlertDismissAjaxView.as_view(),
+        name='ajax_dismiss',
+    ),
+]