+++ config/urls.py
@@ -43,7 +43,9 @@ urlpatterns = [
     # User Profile URLs
     url(r'^user/', include('userprofile.urls')),
     # Admin Alerts URLs
-    url(r'^alerts/', include('adminalerts.urls')),
+    url(r'^alerts/admin/', include('adminalerts.urls')),
+    # App Alerts URLs
+    url(r'^alerts/app/', include('appalerts.urls')),
     # Site Info URLs
     url(r'^siteinfo/', include('siteinfo.urls')),
     # API Tokens URLs