+++ CHANGELOG.rst
@@ -15,6 +15,8 @@ Added
 
 - **Adminalerts**
     - ``get_statistics()`` implementation
+- **Appalerts**
+    - Add site app and backend for app alerts (#642)
 - **Projectroles**
     - ``geticons`` management command for retrieving Iconify icons (#54)
     - ``spin`` class in ``projectroles.css`` for spinning icon support (#54)
@@ -50,6 +52,7 @@ Changed
     - Set parent owner as initial owner in project form (#667)
     - Always show Django admin warning (#677)
     - Modify signature of ``get_history_dropdown()`` template tag (#668)
+    - Add default ``superuser`` value to ``LiveUserMixin._make_user()``
 - **Siteinfo**
     - Tabbed layout in site info view
 - **Timeline**